+++ packages/components/src/HeaderBar/HeaderBar.stories.js
@@ -1,12 +1,13 @@
 import { action } from '@storybook/addon-actions';
-import { makeDecorator } from '@storybook/addons';
+import Immutable from 'immutable';
 
-import Immutable from 'immutable'; // eslint-disable-line import/no-extraneous-dependencies
+import assetsApi from '@talend/assets-api';
+import tokens from '@talend/design-tokens';
 
+import AppSwitcher from '../AppSwitcher';
+// eslint-disable-line import/no-extraneous-dependencies
 import Icon from '../Icon';
 import HeaderBar from './HeaderBar.component';
-import AppSwitcher from '../AppSwitcher';
-import assetsApi from '@talend/assets-api';
 
 const props = {
 	brand: {
@@ -40,17 +41,17 @@ const props = {
 	products: {
 		items: [
 			{
-				icon: 'talend-tdp-colored',
+				icon: 'talend-tdp-negative',
 				key: 'tdp',
 				label: 'Data Preparation',
 			},
 			{
-				icon: 'talend-tic-colored',
+				icon: 'talend-tic-negative',
 				key: 'tic',
 				label: 'Integration Cloud',
 			},
 			{
-				icon: 'talend-tmc-colored',
+				icon: 'talend-tmc-negative',
 				key: 'tmc',
 				label: 'Management Console',
 			},
@@ -96,6 +97,11 @@ function IntercomComponent() {
 
 export default {
 	title: 'Components/Navigation/HeaderBar',
+	decorators: Story => (
+		<div style={{ background: tokens.coralColorBrandBackgroundInvertedActive, height: '100vh' }}>
+			<Story />
+		</div>
+	),
 };
 
 export const Default = () => {
@@ -138,7 +144,7 @@ export const WithBrandIcon = () => {
 		...props,
 		brand: {
 			...props.brand,
-			icon: 'talend-tmc-positive',
+			icon: 'talend-tmc-negative',
 		},
 	}).toJS();
 	return <HeaderBar {...headerProps} />;