+++ packages/storybook-docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @talend/storybook-docs
 
+## 2.3.0
+
+### Minor Changes
+
+- b2d93a4: feat: add @qlik-light theme
+
+### Patch Changes
+
+- Updated dependencies [b2d93a4]
+  - @talend/design-tokens@2.11.0
+
 ## 2.2.1
 
 ### Patch Changes