+++ tools/scripts-config-storybook-lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @talend/scripts-config-storybook-lib
 
+## 5.3.0
+
+### Minor Changes
+
+- b2d93a4: feat: add @qlik-light theme
+
 ## 5.2.1
 
 ### Patch Changes